+++ examples/general/SRL/lib/regex_repr.rb
@@ -1,0 +1,5 @@
+require_relative './regex/character'
+require_relative './regex/char_range'
+require_relative './regex/multiplicity'
+require_relative './regex/repetition'
+require_relative './regex/char_class'