+++ examples/general/SRL/spec/spec_helper.rb
@@ -1,0 +1,25 @@
+# File: spec_helper.rb
+# Purpose: utility file that is loaded by all our RSpec files
+
+require 'simplecov'
+
+SimpleCov.formatter = SimpleCov::Formatter::MultiFormatter.new(
+  [
+    SimpleCov::Formatter::HTMLFormatter,
+  ]
+)
+
+require 'pp'    # Use pretty-print for debugging purposes
+require 'rspec' # Use the RSpec framework
+
+RSpec.configure do |config|
+  config.expect_with :rspec do |c|
+    # Disable the `should` syntax...
+    c.syntax = :expect
+  end
+
+  # Display stack trace in case of failure
+  config.full_backtrace = true
+end
+
+# End of file